use params to select DB rather than store
