+++ packages/builder/src/components/nav/DatabasesList.svelte
@@ -1,5 +1,6 @@
 <script>
   import { tick, onMount } from "svelte"
+  import { goto } from "@sveltech/routify"
   import { store, backendUiStore } from "builderStore"
   import api from "builderStore/api"
   import getIcon from "../common/icon"
@@ -44,7 +45,9 @@
         </span>
         <button
           class:active={database.id === $backendUiStore.selectedDatabase.id}
-          on:click={() => selectDatabase(database)}>
+          on:click={() => {
+            $goto(`./database/${database.id}`), selectDatabase(database)
+          }}>
           {database.name}
         </button>
         <i
