+++ packages/builder/src/pages/[application]/backend/database/[selectedDatabase]/index.svelte
@@ -1,0 +1,43 @@
+<script>
+  import ModelDataTable from "components/database/ModelDataTable"
+  import { store, backendUiStore } from "builderStore"
+  import ActionButton from "components/common/ActionButton.svelte"
+  import * as api from "components/database/ModelDataTable/api"
+
+  export let selectedDatabase
+
+  let selectedRecord
+
+  async function selectRecord(record) {
+    selectedRecord = await api.loadRecord(record.key, {
+      appname: $store.appname,
+      instanceId: selectedDatabase,
+    })
+  }
+
+  $: breadcrumbs = $backendUiStore.breadcrumbs.join(" / ")
+</script>
+
+<div class="database-actions">
+  <div class="budibase__label--big">{breadcrumbs}</div>
+  {#if $backendUiStore.selectedDatabase.id}
+    <ActionButton
+      primary
+      on:click={() => {
+        selectedRecord = null
+        backendUiStore.actions.modals.show('RECORD')
+      }}>
+      Create new record
+    </ActionButton>
+  {/if}
+</div>
+{#if $backendUiStore.selectedDatabase.id}
+  <ModelDataTable {selectRecord} />
+{:else}Please select a database{/if}
+
+<style>
+  .database-actions {
+    display: flex;
+    justify-content: space-between;
+  }
+</style>
